[KC][TLS] Disabled TLS for dev KC container
Enabling TLS on the Keycloak container with self-signed certificates for testing & dev purposes proves to be a pain.
Moreover, in real life, we would be either using real TLS certs, or connecting to other OAUth2 providers (e.g. Google, Github, etc, coming in the future), or implementing it in a more serious way (to be investigated & defined).
So for now, we will be considering the local KC as  a "provided" 3rd party tool, using it just as it comes.